11_4 update test heat beat in qt exe

+++ AntennaTracker/GCS_Mavlink.cpp
@@ -406,7 +406,7 @@ void GCS_MAVLINK_Tracker::mavlink_check_target(const mavlink_message_t &msg)
     tracker.target_set = true;
 }
 
-uint8_t GCS_MAVLINK_Tracker::sysid_my_gcs() const
+uint16_t GCS_MAVLINK_Tracker::sysid_my_gcs() const
 {
     return tracker.g.sysid_my_gcs;
 }
