+++ AntennaTracker/GCS_Mavlink.h
@@ -9,7 +9,7 @@ public:
 
     using GCS_MAVLINK::GCS_MAVLINK;
 
-    uint8_t sysid_my_gcs() const override;
+    uint16_t sysid_my_gcs() const override;
 
 protected:
