+++ ArduCopter/GCS_Copter.cpp
@@ -2,7 +2,7 @@
 
 #include "Copter.h"
 
-uint8_t GCS_Copter::sysid_this_mav() const
+uint16_t GCS_Copter::sysid_this_mav() const
 {
     return copter.g.sysid_this_mav;
 }
