+++ ArduCopter/GCS_Copter.h
@@ -28,7 +28,7 @@ public:
     bool simple_input_active() const override;
     bool supersimple_input_active() const override;
 
-    uint8_t sysid_this_mav() const override;
+    uint16_t sysid_this_mav() const override;
 
 protected:
